requirements.txt
requests==2.25.1
selenium==3.141.0
pytest==6.2.4
Appium_Python_Client==1.2.0
